requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
filelock==3.14.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.1
idna==3.7
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
PyYAML==6.0.1
requests==2.32.2
sentry-sdk==2.2.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
timm==0.6.11
torch==1.13.0
torchvision==0.14.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
wandb==0.17.0
